We now have a separate signed-contract workflow for approved customers, and only Admin or platform Manager can manage it. What changed: added backend upload/delete endpoints for signed contracts stored them separately from lender contracts in facility_signed_contract_documents exposed them in the lending portfolio API added a new UI section in the approved customer file:العقود الموقعة upload signed contracts download signed contracts delete signed contracts
Updated files:
[CreditFacilitySignedContractRequest.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/app/Http/Requests/CreditFacilitySignedContractRequest.php)
[CreditFacility.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/app/Models/CreditFacility.php)
[CreditFacilityResource.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/app/Http/Resources/CreditFacilityResource.php)
[CreditApplicationService.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/app/Services/CreditApplicationService.php)
[CreditApplicationController.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/app/Http/Controllers/Admin/CreditApplicationController.php)
[api.php](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/routes/api.php)
[creditApplicationReview.js](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/resources/js/store/modules/creditApplicationReview.js)
[LendingPortfolioShowComponent.vue](/Users/minaelhamy/Downloads/Starcom codebase/starcom web source code/starcom webstore/resources/js/components/admin/lendingPortfolio/LendingPortfolioShowComponent.vue)

+++ starcom webstore/app/Services/CreditApplicationService.php
@@ -12,6 +12,7 @@ use App\Http\Requests\CreditApplicationNoteRequest;
 use App\Http\Requests\CreditFacilityAssignmentRequest;
 use App\Http\Requests\CreditFacilityContractRequest;
 use App\Http\Requests\CreditFacilityDatesRequest;
+use App\Http\Requests\CreditFacilitySignedContractRequest;
 use App\Http\Requests\CreditApplicationStoreRequest;
 use App\Http\Requests\CreditApplicationUpdateRequest;
 use App\Http\Requests\PaginateRequest;
@@ -588,6 +589,43 @@ class CreditApplicationService
         }
     }
 
+    public function uploadSignedFacilityContracts(CreditFacility $creditFacility, CreditFacilitySignedContractRequest $request): CreditFacility
+    {
+        try {
+            $actor = Auth::user();
+
+            if (
+                !$actor->hasRole(EnumRole::ADMIN) &&
+                !$actor->hasRole(EnumRole::MANAGER)
+            ) {
+                throw new Exception(trans('all.message.permission_denied'), 422);
+            }
+
+            if ($creditFacility->status !== CreditFacilityStatus::APPROVED) {
+                throw new Exception('يمكن رفع العقود الموقعة فقط بعد اعتماد التمويل.', 422);
+            }
+
+            foreach ($request->file('signed_contract_documents', []) as $contractDocument) {
+                $creditFacility->addMedia($contractDocument)->toMediaCollection('facility_signed_contract_documents');
+            }
+
+            return $creditFacility->load([
+                'user',
+                'user.latestAddress',
+                'application.user',
+                'application.notesHistory.author.financialInstitutionOwner.financialInstitutionProfile',
+                'notesHistory.author.financialInstitutionOwner.financialInstitutionProfile',
+                'application.facilities.institution.financialInstitutionProfile',
+                'application.facilities.employee',
+                'institution.financialInstitutionProfile',
+                'employee',
+            ]);
+        } catch (Exception $exception) {
+            Log::info($exception->getMessage());
+            throw new Exception(QueryExceptionLibrary::message($exception), 422);
+        }
+    }
+
     public function deleteFacilityContract(CreditFacility $creditFacility, int $mediaId): CreditFacility
     {
         try {
@@ -617,6 +655,45 @@ class CreditApplicationService
 
             if (!$media instanceof Media) {
                 throw new Exception('العقد المطلوب غير موجود.', 422);
+            }
+
+            $media->delete();
+
+            return $creditFacility->load([
+                'user',
+                'user.latestAddress',
+                'application.user',
+                'application.notesHistory.author.financialInstitutionOwner.financialInstitutionProfile',
+                'notesHistory.author.financialInstitutionOwner.financialInstitutionProfile',
+                'application.facilities.institution.financialInstitutionProfile',
+                'application.facilities.employee',
+                'institution.financialInstitutionProfile',
+                'employee',
+            ]);
+        } catch (Exception $exception) {
+            Log::info($exception->getMessage());
+            throw new Exception(QueryExceptionLibrary::message($exception), 422);
+        }
+    }
+
+    public function deleteSignedFacilityContract(CreditFacility $creditFacility, int $mediaId): CreditFacility
+    {
+        try {
+            $actor = Auth::user();
+
+            if (
+                !$actor->hasRole(EnumRole::ADMIN) &&
+                !$actor->hasRole(EnumRole::MANAGER)
+            ) {
+                throw new Exception(trans('all.message.permission_denied'), 422);
+            }
+
+            $media = $creditFacility
+                ->getMedia('facility_signed_contract_documents')
+                ->firstWhere('id', (int) $mediaId);
+
+            if (!$media instanceof Media) {
+                throw new Exception('العقد الموقع المطلوب غير موجود.', 422);
             }
 
             $media->delete();
